# Optimize computeScore